+++ res/raw/changelog.md
@@ -1,7 +1,7 @@
 ## 4.57 (may/june 2025)
 
 - german translation (Thanks to Ralph Bieber)
-- You can now use and URL from which to fetch settings. 
+- You can now use a URL from which to fetch settings. 
   - Typically used to more easily have all devices for a certain event or league use the same settings.
 
 ## 4.56 (april 2025)